Fix WebSocketServerExamplesTest to work with surefire update
