+++ jetty-websocket/websocket-javax-tests/src/test/resources/jetty-websocket-httpclient.xml
